fix: moved nao_core/conftest.py logic to nao_core/commands/conftest.py
